backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.0.3
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
authlib==1.2.1
redis==5.0.1
aioredis==2.0.1
fastapi-limiter==0.1.6
psycopg2-binary==2.9.9
email-validator>=2.0.0